java/test/processing/mode/java/preproc/util/strategy/MissingIdentifierMessageSimplifierStrategyTest.java
package processing.mode.java.preproc.util.strategy;

import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;

import java.util.Optional;

import static org.junit.Assert.*;

public class MissingIdentifierMessageSimplifierStrategyTest {

  private MissingIdentifierMessageSimplifierStrategy strategy;

  @Before
  public void setup() {
    strategy = new MissingIdentifierMessageSimplifierStrategy();
  }

  @Test
  public void testPresent() {
    Optional<String> msg = strategy.simplify("Missing identifier at ';'");
    Assert.assertTrue(msg.isPresent());
  }

  @Test
  public void testNotPresent() {
    Optional<String> msg = strategy.simplify("String x = \" \\\" \"");
    Assert.assertTrue(msg.isEmpty());
  }

}